Break apart segments modules
