+++ src/sqlfluff/core/parser/segments/__init__.py
@@ -1,0 +1,16 @@
+"""Definitions of the segment classes."""
+
+# flake8: noqa: F401
+
+from .base import BaseSegment, UnparsableSegment
+from .generator import SegmentGenerator
+from .raw import RawSegment
+from .ephemeral import EphemeralSegment
+from .indent import Indent, Dedent
+from .keyword import (
+    KeywordSegment,
+    SymbolSegment,
+    ReSegment,
+    NamedSegment,
+    LambdaSegment,
+)
