+++ src/sqlfluff/core/parser/segments/generator.py
