+++ test/core/parser/segments_base_test.py
@@ -8,7 +8,7 @@ from sqlfluff.core.parser import (
     RawSegment,
     BaseSegment,
 )
-from sqlfluff.core.parser.segments_base import trim_non_code
+from sqlfluff.core.parser.helpers import trim_non_code
 from sqlfluff.core.dialects import ansi_dialect
